backend/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.2
Werkzeug==2.3.6
SQLAlchemy==2.0.19
python-dotenv==1.0.0

# Google AI
google-generativeai==0.3.1
google-cloud-vision==3.4.4
google-auth==2.22.0

# Computer Vision and Face Recognition
opencv-python==4.8.0.74
numpy==1.24.3
Pillow==10.0.0

# Advanced Face Recognition (RECOMMENDED for high accuracy)
# Install these for ArcFace recognition:
# pip install insightface onnxruntime
# If installation fails, system will use fallback mode

# Optional: Simple face recognition (fallback)
# face-recognition==1.3.0
# dlib==19.24.2

# NLP and Text Processing
nltk==3.8.1
pdfplumber==0.10.2
python-docx==0.8.11
PyMuPDF==1.23.5  # For handwriting PDF processing

# Optional: Advanced ML features
transformers==4.31.0
torch==2.0.1
sympy==1.12

# Utilities
requests==2.31.0
